Change how export labels are generated
Ticket: #180
